Commit 4

+++ routing_settings.h
@@ -2,6 +2,6 @@
 #include <cstdint>
 
 struct RoutingSettings {
-	uint32_t bus_wait_time; //В минутах
+	double bus_wait_time; //В минутах
 	double bus_velocity; //В км/час
 };
